fix(plugins/plugin-kubectl): watcher table disappears when kui is launched as kubectl plugin
Fixes #4120
