+++ .travis.yml
@@ -52,7 +52,7 @@ jobs:
     - if: os = windows AND env(NO_WINDOWS) IS present
 
   include:
-    - env: LAYERS="k8s helm logs" MOCHA_RUN_TARGET=electron NEEDS_K8S=true NEEDS_HELM=true
+    - env: LAYERS="k8s k8s-popup helm logs" MOCHA_RUN_TARGET=electron NEEDS_K8S=true NEEDS_HELM=true
     - env: LAYERS=k8s1 MOCHA_RUN_TARGET=webpack  NEEDS_K8S=true NEEDS_OC=true
     - env: LAYERS=k8s2 MOCHA_RUN_TARGET=webpack  NEEDS_K8S=true NEEDS_OC=true   NEEDS_TOP=true
     - env: N="API" CLIENT="test" MOCHA_RUN_TARGET="webpack" LAYERS="response" CODECOV_OF_PRESCAN=true
