+++ plugins/plugin-kubectl/src/test/k8s-popup/dash-dash-ui.ts
