+++ plugins/plugin-kubectl/src/test/k8s-popup/headless-create-pod.ts
